+++ source/Server.Contracts/Accounts/IAccountDetails.cs
@@ -1,4 +1,4 @@
-namespace Sashimi.Accounts.Accounts
+namespace Sashimi.Server.Contracts.Accounts
 {
     public interface IAccountDetails
     {
