+++ source/Server.Contracts/Accounts/Temp/AmazonWebServicesAccountDetails.cs
@@ -1,6 +1,6 @@
 using Octopus.Data.Model;
 
-namespace Sashimi.Accounts.Accounts
+namespace Sashimi.Server.Contracts.Accounts.Temp
 {
     public class AmazonWebServicesAccountDetails : IAccountDetails
     {
